pyproject.toml
[tool.poetry]
name = "microcurrency"
version = "0.1.0"
description = ""
authors = ["BadPythonCoder <nuh@uh.uh>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.1"

[tool.poetry.scripts]
bot = "microcurrency.__init__:start"
create_money = "microcurrency.scripts:create_money"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
